Auth checks when user visits the page.

+++ routes/api.php
@@ -7,8 +7,8 @@ use Illuminate\Support\Facades\Route;
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
-Route::post('/create-ad', [ExampleController::class, 'storeAd']);
-Route::get('/get-ads', [ExampleController::class, 'getAds']);
-Route::delete('/remove-ad/{id}', [ExampleController::class, 'removeAd']);
+Route::post('/create-ad', [ExampleController::class, 'storeAd'])->middleware(['rolepermission']);
+Route::get('/get-ads', [ExampleController::class, 'getAds'])->middleware(['rolepermission']);
+Route::delete('/remove-ad/{id}', [ExampleController::class, 'removeAd'])->middleware(['rolepermission']);
